Moved all indirectly executed test classes from nunit.testdata to their corresponding test fixtures and marking them as explicit.
