+++ src/NUnitFramework/testdata/ExecutionHookTests/ExecutionHooksTestData.cs
@@ -252,409 +252,6 @@ namespace NUnit.TestData.ExecutionHookTests
         }
     }
 
-    public class FailingTestWithTestHookOnMethod
-    {
-        [OneTimeSetUp]
-        public void OneTimeSetUp()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [OneTimeTearDown]
-        public void OneTimeTearDown()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [SetUp]
-        public void SetUp()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [TearDown]
-        public void TearDown()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [Test]
-        [ActivateBeforeTestHook]
-        [ActivateAfterTestHook]
-        public void EmptyTest()
-        {
-            TestLog.LogCurrentMethod();
-            Assert.Fail("Some failure in test");
-        }
-    }
-
-    public class TestWithTestHooksOnMethodAndErrorOnAfterTestHook
-    {
-        [OneTimeSetUp]
-        public void OneTimeSetUp()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [OneTimeTearDown]
-        public void OneTimeTearDown()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [SetUp]
-        public void SetUp()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [TearDown]
-        public void TearDown()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [Test]
-        [ActivateBeforeTestHook]
-        [ActivateAfterTestHookThrowingException]
-        public void EmptyTest()
-        {
-            TestLog.LogCurrentMethod();
-        }
-    }
-
-    public class TestWithTestHooksOnMethodAndErrorOnBeforeTestHook
-    {
-        [OneTimeSetUp]
-        public void OneTimeSetUp()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [OneTimeTearDown]
-        public void OneTimeTearDown()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [SetUp]
-        public void SetUp()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [TearDown]
-        public void TearDown()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [Test]
-        [ActivateBeforeTestHookThrowingException]
-        [ActivateAfterTestHook]
-        public void EmptyTest()
-        {
-            TestLog.LogCurrentMethod();
-        }
-    }
-
-    public class TestWithTestHooksOnMethod
-    {
-        [OneTimeSetUp]
-        public void OneTimeSetUp()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [OneTimeTearDown]
-        public void OneTimeTearDown()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [SetUp]
-        public void SetUp()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [TearDown]
-        public void TearDown()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [Test]
-        [ActivateBeforeTestHook]
-        [ActivateAfterTestHook]
-        public void EmptyTest()
-        {
-            TestLog.LogCurrentMethod();
-        }
-    }
-
-    public class TestWithAfterTestHookOnMethod
-    {
-        [OneTimeSetUp]
-        public void OneTimeSetUp()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [OneTimeTearDown]
-        public void OneTimeTearDown()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [SetUp]
-        public void SetUp()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [TearDown]
-        public void TearDown()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [Test]
-        [ActivateAfterTestHook]
-        public void EmptyTest()
-        {
-            TestLog.LogCurrentMethod();
-        }
-    }
-
-    public class TestWithBeforeTestHookOnMethod
-    {
-        [OneTimeSetUp]
-        public void OneTimeSetUp()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [OneTimeTearDown]
-        public void OneTimeTearDown()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [SetUp]
-        public void SetUp()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [TearDown]
-        public void TearDown()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [Test]
-        [ActivateBeforeTestHook]
-        public void EmptyTest()
-        {
-            TestLog.LogCurrentMethod();
-        }
-    }
-
-    public class TestWithNormalAndLongRunningTestHooks
-    {
-        [OneTimeSetUp]
-        public void OneTimeSetUp()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [OneTimeTearDown]
-        public void OneTimeTearDown()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [SetUp]
-        public void SetUp()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [TearDown]
-        public void TearDown()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [Test]
-        [ActivateBeforeTestHook]
-        [ActivateLongRunningBeforeTestHook]
-        [ActivateAfterTestHook]
-        [ActivateLongRunningAfterTestHook]
-        public void EmptyTest()
-        {
-            TestLog.LogCurrentMethod();
-        }
-    }
-
-    [ActivateClassLevelBeforeTestHooks]
-    [ActivateClassLevelAfterTestHooks]
-    public class TestClassWithTestHooksOneTestWithoutAndOneWithMethodTestHooks
-    {
-        [OneTimeSetUp]
-        public void OneTimeSetUp()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [OneTimeTearDown]
-        public void OneTimeTearDown()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [SetUp]
-        public void SetUp()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [TearDown]
-        public void TearDown()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [Test]
-        [ActivateMethodLevelBeforeTestHooks]
-        [ActivateMethodLevelAfterTestHooks]
-        public void EmptyTestWithHooks()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [Test]
-        public void EmptyTestWithoutHooks()
-        {
-            TestLog.LogCurrentMethod();
-        }
-    }
-
-    [SomeTestAction]
-    public class TestWithTestHooksAndClassTestActionAttribute
-    {
-        [OneTimeSetUp]
-        public void OneTimeSetUp()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [OneTimeTearDown]
-        public void OneTimeTearDown()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [SetUp]
-        public void SetUp()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [TearDown]
-        public void TearDown()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [Test]
-        [ActivateBeforeTestHook]
-        [ActivateAfterTestHook]
-        public void EmptyTest()
-        {
-            TestLog.LogCurrentMethod();
-        }
-    }
-
-    [TestFixture]
-    [LogTestAction]
-    [TestActionLoggingExecutionHooks]
-    public class TestClassWithTestAction
-    {
-        [Test]
-        public void TestUnderTest()
-        {
-            TestLog.LogCurrentMethod();
-        }
-    }
-
-    /// <remarks>
-    /// This TestUnderTest class is failing by intention. Therefore is currently cannot reside in the same
-    /// class as the calling test inside nunit.framework.tests since it would fail the whole test suite.
-    /// Once https://github.com/nunit/nunit/issues/5002 is solved it can be moved back and marked as explicit.
-    /// Then it can be executed with the TestBuilder while still not execute during the test suite execution.
-    /// </remarks>
-    public class ExecutionProceedsOnlyAfterAllAfterTestHooksExecute_TestUnderTest
-    {
-        [Test]
-        [ActivateAfterTestHook]
-        [ActivateAfterTestHook]
-        [ActivateAfterTestHook]
-        [ActivateAfterTestHookThrowingException]
-        public void TestPasses()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [TearDown]
-        public void TearDown()
-        {
-            TestLog.LogCurrentMethod();
-        }
-
-        [OneTimeTearDown]
-        public void OneTimeTearDown()
-        {
-            TestLog.LogCurrentMethod();
-        }
-    }
-
-    /// <remarks>
-    /// This TestUnderTest class is failing by intention. Therefore is currently cannot reside in the same
-    /// class as the calling test inside nunit.framework.tests since it would fail the whole test suite.
-    /// Once https://github.com/nunit/nunit/issues/5002 is solved it can be moved back and marked as explicit.
-    /// Then it can be executed with the TestBuilder while still not execute during the test suite execution.
-    /// </remarks>
-    public class SynchronousHookInvocationTests_TestUnderTest
-    {
-        [SetUp]
-        public void Setup()
-        {
-            TestExecutionContext.CurrentContext
-                                .CurrentTest.Properties
-                                .Add("TestThreadId", Environment.CurrentManagedThreadId);
-        }
-
-        [Test, ActivateSynchronousTestHook]
-        public void TestPasses_WithAssertPass()
-        {
-            Assert.Pass("Test passed.");
-        }
-
-        [Test, ActivateSynchronousTestHook]
-        public void TestFails_WithAssertFail()
-        {
-            Assert.Fail("Test failed with Assert.Fail");
-        }
-
-        [Test, ActivateSynchronousTestHook]
-        public void TestFails_WithException()
-        {
-            throw new Exception("Test failed with Exception");
-        }
-    }
-
     public static class HookIdentifiers
     {
         public static readonly string Hook = "_Hook";
