+++ src/NUnitFramework/tests/ExecutionHooks/Execution/AfterTestHookTests.cs
@@ -1,5 +1,6 @@
 // Copyright (c) Charlie Poole, Rob Prouse and Contributors. MIT License - see LICENSE.txt
 
+using NUnit.Framework.Internal;
 using NUnit.Framework.Tests.TestUtilities;
 using NUnit.TestData.ExecutionHookTests;
 
@@ -7,12 +8,47 @@ namespace NUnit.Framework.Tests.ExecutionHooks.Execution
 {
     internal class AfterTestHookTests
     {
+        [Explicit($"This test should only be run as part of the {nameof(ExecutionProceedsAfterTheAfterTestHookCompletes)} test")]
+        public class TestWithAfterTestHookOnMethod
+        {
+            [OneTimeSetUp]
+            public void OneTimeSetUp()
+            {
+                TestLog.LogCurrentMethod();
+            }
+
+            [OneTimeTearDown]
+            public void OneTimeTearDown()
+            {
+                TestLog.LogCurrentMethod();
+            }
+
+            [SetUp]
+            public void SetUp()
+            {
+                TestLog.LogCurrentMethod();
+            }
+
+            [TearDown]
+            public void TearDown()
+            {
+                TestLog.LogCurrentMethod();
+            }
+
+            [Test]
+            [ActivateAfterTestHook]
+            public void EmptyTest()
+            {
+                TestLog.LogCurrentMethod();
+            }
+        }
+
         [Test]
         public void ExecutionProceedsAfterTheAfterTestHookCompletes()
         {
             TestLog.Clear();
 
-            var workItem = TestBuilder.CreateWorkItem(typeof(TestWithAfterTestHookOnMethod));
+            var workItem = TestBuilder.CreateWorkItem(typeof(TestWithAfterTestHookOnMethod), TestFilter.Explicit);
             workItem.Execute();
 
             Assert.That(TestLog.Logs, Is.EqualTo([
