+++ src/NUnitFramework/tests/ExecutionHooks/Execution/BeforeTestHookTests.cs
@@ -1,5 +1,6 @@
 // Copyright (c) Charlie Poole, Rob Prouse and Contributors. MIT License - see LICENSE.txt
 
+using NUnit.Framework.Internal;
 using NUnit.Framework.Tests.TestUtilities;
 using NUnit.TestData.ExecutionHookTests;
 
@@ -7,12 +8,47 @@ namespace NUnit.Framework.Tests.ExecutionHooks.Execution
 {
     internal class BeforeTestHookTests
     {
+        [Explicit($"This test should only be run as part of the {nameof(ExecutionProceedsAfterBeforeTestHookCompletes)} test")]
+        public class TestWithBeforeTestHookOnMethod
+        {
+            [OneTimeSetUp]
+            public void OneTimeSetUp()
+            {
+                TestLog.LogCurrentMethod();
+            }
+
+            [OneTimeTearDown]
+            public void OneTimeTearDown()
+            {
+                TestLog.LogCurrentMethod();
+            }
+
+            [SetUp]
+            public void SetUp()
+            {
+                TestLog.LogCurrentMethod();
+            }
+
+            [TearDown]
+            public void TearDown()
+            {
+                TestLog.LogCurrentMethod();
+            }
+
+            [Test]
+            [ActivateBeforeTestHook]
+            public void EmptyTest()
+            {
+                TestLog.LogCurrentMethod();
+            }
+        }
+
         [Test]
         public void ExecutionProceedsAfterBeforeTestHookCompletes()
         {
             TestLog.Clear();
 
-            var workItem = TestBuilder.CreateWorkItem(typeof(TestWithBeforeTestHookOnMethod));
+            var workItem = TestBuilder.CreateWorkItem(typeof(TestWithBeforeTestHookOnMethod), TestFilter.Explicit);
             workItem.Execute();
 
             Assert.That(TestLog.Logs, Is.EqualTo([
