+++ src/NUnitFramework/tests/ExecutionHooks/Execution/CombinedHookAtClassAndMethodLevelTests.cs
@@ -1,5 +1,6 @@
 // Copyright (c) Charlie Poole, Rob Prouse and Contributors. MIT License - see LICENSE.txt
 
+using NUnit.Framework.Internal;
 using NUnit.Framework.Tests.TestUtilities;
 using NUnit.TestData.ExecutionHookTests;
 
@@ -7,12 +8,56 @@ namespace NUnit.Framework.Tests.ExecutionHooks.Execution
 {
     internal class CombinedHookAtClassAndMethodLevelTests
     {
+        [Explicit($"This test should only be run as part of the {nameof(ExecutionProceedsAfterBothTestHookCompletes)} test")]
+        [ActivateClassLevelBeforeTestHooks]
+        [ActivateClassLevelAfterTestHooks]
+        public class TestClassWithTestHooksOneTestWithoutAndOneWithMethodTestHooks
+        {
+            [OneTimeSetUp]
+            public void OneTimeSetUp()
+            {
+                TestLog.LogCurrentMethod();
+            }
+
+            [OneTimeTearDown]
+            public void OneTimeTearDown()
+            {
+                TestLog.LogCurrentMethod();
+            }
+
+            [SetUp]
+            public void SetUp()
+            {
+                TestLog.LogCurrentMethod();
+            }
+
+            [TearDown]
+            public void TearDown()
+            {
+                TestLog.LogCurrentMethod();
+            }
+
+            [Test]
+            [ActivateMethodLevelBeforeTestHooks]
+            [ActivateMethodLevelAfterTestHooks]
+            public void EmptyTestWithHooks()
+            {
+                TestLog.LogCurrentMethod();
+            }
+
+            [Test]
+            public void EmptyTestWithoutHooks()
+            {
+                TestLog.LogCurrentMethod();
+            }
+        }
+
         [Test]
         public void ExecutionProceedsAfterBothTestHookCompletes()
         {
             TestLog.Clear();
 
-            var workItem = TestBuilder.CreateWorkItem(typeof(TestClassWithTestHooksOneTestWithoutAndOneWithMethodTestHooks));
+            var workItem = TestBuilder.CreateWorkItem(typeof(TestClassWithTestHooksOneTestWithoutAndOneWithMethodTestHooks), TestFilter.Explicit);
             workItem.Execute();
 
             Assert.That(TestLog.Logs, Is.EqualTo([
