+++ src/NUnitFramework/tests/ExecutionHooks/ExecutionSequence/ExecutionProceedsOnlyAfterAllAfterTestHooksExecute.cs
@@ -1,5 +1,6 @@
 // Copyright (c) Charlie Poole, Rob Prouse and Contributors. MIT License - see LICENSE.txt
 
+using NUnit.Framework.Internal;
 using NUnit.Framework.Tests.TestUtilities;
 using NUnit.TestData.ExecutionHookTests;
 
@@ -7,12 +8,38 @@ namespace NUnit.Framework.Tests.ExecutionHooks.ExecutionSequence
 {
     internal class ExecutionProceedsOnlyAfterAllAfterTestHooksExecute
     {
+        [Explicit($"This test should only be run as part of the {nameof(TestProceedsAfterAllAfterTestHooksExecute)} test")]
+        public class ExecutionProceedsOnlyAfterAllAfterTestHooksExecute_TestUnderTest
+        {
+            [Test]
+            [ActivateAfterTestHook]
+            [ActivateAfterTestHook]
+            [ActivateAfterTestHook]
+            [ActivateAfterTestHookThrowingException]
+            public void TestPasses()
+            {
+                TestLog.LogCurrentMethod();
+            }
+
+            [TearDown]
+            public void TearDown()
+            {
+                TestLog.LogCurrentMethod();
+            }
+
+            [OneTimeTearDown]
+            public void OneTimeTearDown()
+            {
+                TestLog.LogCurrentMethod();
+            }
+        }
+
         [Test]
         public void TestProceedsAfterAllAfterTestHooksExecute()
         {
             TestLog.Clear();
 
-            var workItem = TestBuilder.CreateWorkItem(typeof(ExecutionProceedsOnlyAfterAllAfterTestHooksExecute_TestUnderTest));
+            var workItem = TestBuilder.CreateWorkItem(typeof(ExecutionProceedsOnlyAfterAllAfterTestHooksExecute_TestUnderTest), TestFilter.Explicit);
             workItem.Execute();
 
             Assert.That(TestLog.Logs, Is.EqualTo([
